requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.8
scikit-learn==1.5.2
pypdf==5.9.0
python-docx==1.1.2
pytest==8.3.5
httpx==0.28.1
